Bug 2032985 - Make PBrowser::RequestNativeKeyBindings more idiomatic. r=masayuki
Differential Revision: https://phabricator.services.mozilla.com/D294946

UltraBlame original commit: 8135a9463098a4b7523b9a07ea53d5ded666731e

+++ dom/ipc/BrowserChild.cpp
@@ -2396,8 +2396,7 @@ void BrowserChild::RequestEditCommands(NativeKeyBindingsType aType,
   
   
   WidgetKeyboardEvent localEvent(aEvent);
-  SendRequestNativeKeyBindings(static_cast<uint32_t>(aType), localEvent,
-                               &aCommands);
+  SendRequestNativeKeyBindings(aType, localEvent, &aCommands);
 }
 
 mozilla::ipc::IPCResult BrowserChild::RecvSynthesizedEventResponse(
